add 전체글조회 메소드 구현

+++ board/src/board/Board.java
@@ -16,6 +16,8 @@ public class Board {
 	private final int CREATE_POST = 5;
 	private final int MODIFY_POST = 6;
 	private final int DELETE_POST = 7;
+	private final int ALL_POST = 8;
+	
 	
 	private final int DELETE_USER = 1;
 	private final int CREATE_NOTICE = 2;
@@ -30,7 +32,7 @@ public class Board {
 	private UserManager userManager;
 	private PostManager postManager;
 	private Admin admin;
-	private ArrayList<Post> post;
+	private ArrayList<Post> allPost;
 	
 	private int log;
 	
@@ -42,7 +44,7 @@ public class Board {
 		map = new HashMap<>();
 		userManager = UserManager.getInstance();
 		postManager = PostManager.getInstance();
-		post = AllPost.getInstance();
+		allPost = AllPost.getInstance();
 		admin = Admin.getInstance();
 		map.put(admin, new ArrayList<Post>());
 		log = -1;
@@ -130,8 +132,10 @@ public class Board {
 	private void createPost() {
 		User user = userManager.readUser(log);
 		String name =user.getName();
+		Post post = writePost(name);
 		
-		postManager.createPost(user, writePost(name));
+		allPost.add(post);
+		postManager.createPost(user, post);
 	}
 	
 	private void modifyPost() {
@@ -146,6 +150,8 @@ public class Board {
 		
 		Post post = writePost(user.getName());
 		
+		allPost.add(post);
+		
 		postManager.updatePost(user, post, idx);
 	}
 	
@@ -159,7 +165,53 @@ public class Board {
 			return;
 		}
 		
+		allPost.remove(postManager.readPost(user, idx));
 		postManager.deletePost(user, idx);
+	}
+	
+	private void printPage() {
+		System.out.println("1.전페이지 2.앞페이지 3.나가기");
+	}
+	
+	private int selectPage(int page, int select, int end) {
+		if(select == 1) {
+			if(page == 1)
+				return page;
+			else
+				return page-1;
+		}else if(select == 2) {
+			if(end == allPost.size())
+				return page;
+			else
+				return page+1;
+		}
+		
+		return page;
+	}
+	
+	private void allPost() {
+		int page = 1;
+		
+		while(true) {
+			int start = (page - 1) * 5;
+		    int end = start + 5;
+			if(end >= allPost.size())
+				end = allPost.size();
+			
+			System.out.println("===============");
+			for(int i=start; i<end; i++) {
+				System.out.println(i+1+". " + allPost.get(i).getTitle());
+			}
+			System.out.println("현재 "+page+"페이지");
+			System.out.println("===============");
+
+			printPage();
+			int sel = inputNumber("선택");
+			if(sel == 3)
+				break;
+			
+			page = selectPage(page, sel,end);
+		}
 	}
 	
 	private void printAdminMenu() {
@@ -179,35 +231,32 @@ public class Board {
 		}else if(sel == DELETE_ADMIN_POST) {
 			
 		}else if(sel == LOG_OUT_ADMIN) {
-			
+			logout();
 		}
 	}
 	
 	private void selectMenu() {
 		int sel = inputNumber("선택");
 		
-		if(sel == SIGN_UP) {
+		if(sel == SIGN_UP) 
 			createUser();
-		}else if(sel == LOG_IN) {
+		else if(sel == LOG_IN) 
 			login();
-		}
-		if(log != -1 || log != ADMIN) {	
-			if(sel == LEAVE) {
-				leave();
-			}else if(sel == LOG_OUT) {
-				logout();
-			}else if(sel == CREATE_POST) {
-				createPost();
-			}else if(sel == MODIFY_POST) {
-				modifyPost();
-			}else if(sel == DELETE_POST) {
-				deletePost();
-			}
-		}
+		else if(sel == ALL_POST) 
+			allPost();
 		
-		if(log == ADMIN) {
-			printAdminMenu();
-			seleteAdminMenu();
+		if(log != -1 && log != ADMIN) {	
+			if(sel == LEAVE)
+				leave();
+			else if(sel == LOG_OUT)
+				logout();
+			else if(sel == CREATE_POST)
+				createPost();
+			else if(sel == MODIFY_POST)
+				modifyPost();
+			else if(sel == DELETE_POST)
+				deletePost();
+			
 		}
 		
 	}
@@ -229,6 +278,13 @@ public class Board {
 		while(isRun()) {
 			printMenu();
 			selectMenu();
+			while(log == ADMIN) {
+				if(log == ADMIN) {
+					printAdminMenu();
+					seleteAdminMenu();
+				}
+			}
+			
 
 		}
 	}
